Removed DOM main `subscribe`
