+++ src/editor/responsive-selector/index.js
@@ -4,7 +4,7 @@
 import { __ } from '@wordpress/i18n';
 import { useSelect, useDispatch } from '@wordpress/data';
 import { createBlock } from '@wordpress/blocks';
-import { useRef } from '@wordpress/element';
+import { useCallback, useEffect, useRef } from '@wordpress/element';
 
 /**
  * Internal dependencies
@@ -13,11 +13,14 @@ import Button from '../../components/button';
 import Icon from '../../components/icon';
 import MaxiStyleCardsEditorPopUp from '../style-cards';
 import { setScreenSize } from '../../extensions/styles';
+import { getSiteEditorIframeBody } from '../../extensions/fse';
+import { getPageFonts, loadFonts } from '../../extensions/text/fonts';
 
 /**
  * External dependencies
  */
 import classnames from 'classnames';
+import { isEmpty } from 'lodash';
 
 /**
  * Styles
@@ -103,6 +106,7 @@ const ResponsiveSelector = props => {
 	const settingsRef = useRef(null);
 
 	const { insertBlock } = useDispatch('core/block-editor');
+	const { setMaxiDeviceType } = useDispatch('maxiBlocks');
 
 	const { deviceType, breakpoints, baseBreakpoint } = useSelect(select => {
 		const {
@@ -118,6 +122,220 @@ const ResponsiveSelector = props => {
 			breakpoints: receiveMaxiBreakpoints(),
 			baseBreakpoint,
 		};
+	});
+
+	// TODO: check if it can be reduced to avoid the amount of resources used (MVP)
+	useEffect(() => {
+		// Ensure the editor wrapper has the Maxi responsive attribute
+		const editorWrapper =
+			document.querySelector('.edit-post-visual-editor') ||
+			document.querySelector('.edit-site-visual-editor');
+
+		[editorWrapper, getSiteEditorIframeBody()].forEach(wrapper => {
+			if (
+				wrapper &&
+				!wrapper.getAttributeNames().includes('maxi-blocks-responsive')
+			)
+				wrapper.setAttribute('maxi-blocks-responsive', baseBreakpoint);
+		});
+	});
+
+	// TODO: check if it can be reduced to avoid the amount of resources used (MVP)
+	useEffect(() => {
+		// In case Gutenberg native preview responsive is enabled
+		const tabletPreview = document.querySelector('.is-tablet-preview');
+		const mobilePreview = document.querySelector('.is-mobile-preview');
+
+		if (tabletPreview || mobilePreview) {
+			const previewTarget = tabletPreview ?? mobilePreview;
+
+			const iframe = previewTarget.querySelector(
+				'iframe[name="editor-canvas"]'
+			);
+			const iframeDocument = iframe.contentDocument;
+			const editorWrapper = iframeDocument.body;
+
+			const postEditor = previewTarget.closest(
+				'.edit-post-visual-editor'
+			);
+			const responsiveWidth = postEditor.getAttribute(
+				'maxi-blocks-responsive-width'
+			);
+			const isMaxiPreview = postEditor.getAttribute('is-maxi-preview');
+
+			if (isMaxiPreview) {
+				previewTarget.style.width = `${responsiveWidth}px`;
+				previewTarget.style.boxSizing = 'content-box';
+			}
+
+			if (editorWrapper) {
+				editorWrapper.setAttribute(
+					'maxi-blocks-responsive',
+					tabletPreview ? 's' : 'xs'
+				);
+
+				if (
+					iframe &&
+					!iframeDocument.body.classList.contains(
+						'maxi-blocks--active'
+					)
+				) {
+					// Iframe needs Maxi classes and attributes
+					iframeDocument.body.classList.add('maxi-blocks--active');
+
+					editorWrapper.setAttribute(
+						'maxi-blocks-responsive',
+						tabletPreview ? 's' : 'xs'
+					);
+
+					// Hides scrollbar in firefox
+					iframeDocument.documentElement.style.scrollbarWidth =
+						'none';
+
+					// Copy all fonts to iframe
+					loadFonts(getPageFonts(true), true, iframeDocument);
+
+					// Get all Maxi blocks <style> from <head>
+					// and move to new iframe
+					const maxiStyles = Array.from(
+						document.querySelectorAll('div.maxi-blocks__styles')
+					);
+
+					if (!isEmpty(maxiStyles))
+						maxiStyles.forEach(rawMaxiStyle => {
+							const maxiStyle = rawMaxiStyle.cloneNode(true);
+							const { id } = maxiStyle;
+
+							iframeDocument.querySelector(`#${id}`)?.remove();
+
+							maxiStyle.children[0].innerText =
+								maxiStyle.children[0].innerText.replaceAll(
+									' .edit-post-visual-editor',
+									'.editor-styles-wrapper'
+								);
+
+							iframe.contentDocument.head.appendChild(maxiStyle);
+						});
+
+					// Move Maxi variables to iframe
+					const maxiVariables = document
+						.querySelector('#maxi-blocks-sc-vars-inline-css')
+						?.cloneNode(true);
+
+					if (maxiVariables) {
+						iframeDocument
+							.querySelector('#maxi-blocks-sc-vars-inline-css')
+							?.remove();
+
+						iframe.contentDocument.head.appendChild(maxiVariables);
+					}
+
+					// Ensures all Maxi styles are loaded on iframe
+					const editStyles = iframeDocument.querySelector(
+						'#maxi-blocks-block-editor-css'
+					);
+					const frontStyles = iframeDocument.querySelector(
+						'#maxi-blocks-block-css'
+					);
+
+					if (!editStyles) {
+						const rawEditStyles = document.querySelector(
+							'#maxi-blocks-block-editor-css'
+						);
+
+						iframe.contentDocument.head.appendChild(
+							rawEditStyles.cloneNode(true)
+						);
+					}
+
+					if (!frontStyles) {
+						const rawFrontStyles = document.querySelector(
+							'#maxi-blocks-block-css'
+						);
+
+						iframe.contentDocument.head.appendChild(
+							rawFrontStyles.cloneNode(true)
+						);
+					}
+				}
+			}
+		}
+	});
+
+	// TODO: check if it can be reduced to avoid the amount of resources used (MVP)
+	const onChangeNativeResponsive = useCallback(button => {
+		button.addEventListener('click', e => {
+			const { target } = e;
+			const value = target.innerText;
+			const maxiValue =
+				(value === 'Desktop' && 'general') ||
+				(value === 'Tablet' && 's') ||
+				(value === 'Mobile' && 'xs');
+
+			const editorWrapper =
+				document.querySelector('.edit-post-visual-editor') ||
+				document.querySelector('.edit-site-visual-editor');
+
+			editorWrapper.setAttribute('maxi-blocks-responsive', maxiValue);
+			editorWrapper.removeAttribute('maxi-blocks-responsive-width');
+
+			if (value === 'Desktop') editorWrapper.style.width = '';
+
+			setMaxiDeviceType({
+				deviceType: maxiValue,
+				isGutenbergButton: true,
+			});
+		});
+	});
+
+	useEffect(() => {
+		const previewButton = document.querySelector(
+			'.block-editor-post-preview__button-toggle'
+		);
+
+		if (previewButton) {
+			const config = {
+				attributes: true,
+				childList: false,
+				subtree: false,
+			};
+
+			const callback = mutationsList => {
+				mutationsList.forEach(mutation => {
+					if (mutation.type === 'attributes') {
+						if (mutation.attributeName === 'aria-expanded') {
+							const node = document.querySelector(
+								'.block-editor-post-preview__dropdown-content'
+							);
+							const repeatedNode = document.querySelector(
+								'#maxi-blocks__responsive-toolbar'
+							);
+
+							if (node && !repeatedNode) {
+								// Actions on default responsive values
+								const responsiveButtons = Array.from(
+									node.querySelectorAll(
+										'.block-editor-post-preview__button-resize'
+									)
+								);
+
+								responsiveButtons.forEach(
+									onChangeNativeResponsive
+								);
+							}
+						}
+					}
+				});
+			};
+
+			const observer = new MutationObserver(callback);
+
+			observer.observe(previewButton, config);
+
+			return () => observer.disconnect();
+		}
+
+		return () => {};
 	});
 
 	const addCloudLibrary = () => {
